создан общий компонент для всех страниц
